Revert "feat: 모바일용 헤더 컴포넌트 추가"
This reverts commit e7b3a1858668d99d483f3e2ab3ca5f04f55bc562.
